os2 / emx fixes
