+++ os2/makefile.emx
@@ -29,7 +29,8 @@ CFLAGS_LIB= $(CFLAGS) -DWILDMIDI_BUILD
 CFLAGS_EXE= $(CFLAGS)
 
 OBJ=wm_error.o file_io.o lock.o wildmidi_lib.o reverb.o gus_pat.o f_xmidi.o f_mus.o f_hmp.o f_midi.o f_hmi.o mus2mid.o xmi2mid.o internal_midi.o patches.o sample.o
-PLAYER_OBJ=getopt_long.o wm_tty.o wildmidi.o out_wave.o out_dart.o fileops.o
+PLAYER_OBJ=wildmidi.o getopt_long.o out_wave.o out_dart.o fileops.o
+# wm_tty.o
 
 all: $(LIBSTATIC) $(PLAYER_STATIC)
